refactor: move GameFetcher
